// src/generated/contentful-types.ts
export interface Page {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title: string;
  slug: string;
  contentViewCollection?: { items: Array<ContentView> };
  imageCollection?: { items: Array<{ sys: { id: string }; url?: string }> };
  description?: string;
  sharingTitle?: string;
  sharingImage?: { sys: { id: string }; url?: string };
}

export interface Member {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  firstName?: string;
  lastName?: string;
  email?: string;
  displayName: string;
  category: string;
  displayPriority?: number;
  role: string;
  description?: string;
}
// ...
export interface ContentView {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title?: string;
  type: string;
  dataAdapter: { sys: { id: string } };
  targetCollection?: { items: Array<{ sys: { id: string } }> };
  isAtomic: boolean;
  isTitleVisible: boolean;
  keys?: Array<string>;
  values?: Array<string>;
}

export interface Organization {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title: string;
  entityType: string;
  url?: string;
  logo?: { sys: { id: string }; url?: string };
}

export interface Event {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title: string;
  startDate: string;
  endDate?: string;
  slug: string;
  presenterCollection?: { items: Array<Member> };
  program?: { sys: { id: string } };
  description?: string;
  content?: { json: any };
  imageCollection?: { items: Array<{ sys: { id: string }; url?: string }> };
  callToAction?: { sys: { id: string } };
}
// ...
export interface Program {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title: string;
  description: string;
  url?: string;
  content?: { json: any };
  imageCollection?: { items: Array<{ sys: { id: string }; url?: string }> };
  recurringWeekday?: Array<string>;
}

export interface Post {
  sys: {
    id: string;
    publishedAt?: string;
    updatedAt?: string;
    revision?: number;
  };
  name: string;
  title: string;
  slug: string;
  description?: string;
  content: { json: any };
  date: string;
  imageCollection?: { items: Array<{ sys: { id: string }; url?: string }> };
  videoCollection?: { items: Array<{ sys: { id: string }; url?: string }> };
  callToAction?: { sys: { id: string } };
}